--- src/components/common/Skeleton/Member/index.tsx
@@ -1,13 +1,0 @@
-import * as S from "./style";
-
-const MemberSkeleton = () => {
-  return (
-    <S.MemberSkeletonContainer>
-      {Array.from({ length: 15 }).map((item, idx) => (
-        <S.MemberSkeletonItem key={idx} />
-      ))}
-    </S.MemberSkeletonContainer>
-  );
-};
-
-export default MemberSkeleton;
